pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cvat-ultralytics-bot"
version = "0.1.0"
description = "基于 ultralytics YOLO/SAM 的 CVAT 自动标注工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cvat-sdk==2.40.1",
    "ultralytics>=8.4.0",
    "typer>=0.12.0",
    "Pillow>=10.0.0",
    "numpy>=1.23.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "openai>=1.0.0",
    "modelscope>=1.0.0",
    "addict>=2.4.0",
    "datasets>=3.0.0",
    "oss2>=2.0.0",
    "simplejson>=3.0.0",
    "sortedcontainers>=2.0.0",
]

[project.scripts]
cvat-bot = "cvat_ultralytics_bot.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cvat_ultralytics_bot*"]

[tool.setuptools.package-data]
cvat_ultralytics_bot = ["preset_annotation_configs/*.toml"]
